pyproject.toml
[project]
name = "meu_agente"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "browser-use==0.12.5",
    "langchain-google-genai>=4.2.1",
    "python-dotenv==1.2.1",
]
